refactor: logs-collection is a useLogs hook
